requirements.txt
celery==5.2.1
Django==3.2.9
djangorestframework==3.12.4
djangorestframework_simplejwt==4.8.0
djoser==2.1.0
flower==1.0.0
psycopg2-binary==2.8
redis==4.0.0




